Removed logisland-hadoop-utils and moved classes inside hbase client which use it.
