+++ logisland-components/logisland-services/logisland-service-hbase/logisland-service-hbase_1_1_2-client/src/main/java/com/hurence/logisland/service/hbase/security/KerberosProperties.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hurence.logisland.hadoop;
+package com.hurence.logisland.service.hbase.security;
 
 import com.hurence.logisland.component.PropertyDescriptor;
 import com.hurence.logisland.logging.ComponentLog;
 import com.hurence.logisland.validator.StandardValidators;
-import com.hurence.logisland.validator.ValidationContext;
 import com.hurence.logisland.validator.ValidationResult;
 import com.hurence.logisland.validator.Validator;
 import org.apache.hadoop.conf.Configuration;
