+++ logisland-components/logisland-services/logisland-service-hbase/logisland-service-hbase_1_1_2-client/src/main/java/com/hurence/logisland/service/hbase/security/KerberosTicketRenewer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hurence.logisland.hadoop;
+package com.hurence.logisland.service.hbase.security;
 
 import com.hurence.logisland.logging.ComponentLog;
 import org.apache.hadoop.security.UserGroupInformation;
