+++ logisland-components/logisland-services/logisland-service-hbase/logisland-service-hbase_1_1_2-client/src/test/java/com/hurence/logisland/service/hbase/TestHBase_1_1_2_ClientService.java
@@ -18,12 +18,12 @@ package com.hurence.logisland.service.hbase;
 import com.hurence.logisland.classloading.PluginProxy;
 import com.hurence.logisland.component.InitializationException;
 import com.hurence.logisland.controller.ControllerServiceInitializationContext;
-import com.hurence.logisland.hadoop.KerberosProperties;
 import com.hurence.logisland.service.hbase.put.PutColumn;
 import com.hurence.logisland.service.hbase.put.PutRecord;
 import com.hurence.logisland.service.hbase.scan.Column;
 import com.hurence.logisland.service.hbase.scan.ResultCell;
 import com.hurence.logisland.service.hbase.scan.ResultHandler;
+import com.hurence.logisland.service.hbase.security.KerberosProperties;
 import com.hurence.logisland.util.runner.TestRunner;
 import com.hurence.logisland.util.runner.TestRunners;
 import org.apache.hadoop.hbase.Cell;
