+++ logisland-components/logisland-services/logisland-service-hbase/logisland-service-hbase_1_1_2-client/src/test/java/com/hurence/logisland/service/hbase/security/TestKerberosProperties.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hurence.logisland.hadoop;
+package com.hurence.logisland.service.hbase.security;
 
 import com.hurence.logisland.logging.ComponentLog;
 import com.hurence.logisland.validator.ValidationResult;
