--- logisland-core/logisland-framework/logisland-hadoop-utils/src/test/resources/krb5.conf
@@ -1,12 +1,0 @@
-[libdefaults]
-  default_realm = EXAMPLE.COM
-
-[realms]
-  EXAMPLE.COM = {
-    kdc = kdc1.example.com
-    kdc = kdc2.example.com
-    admin_server = kdc1.example.com
-  }
-
-[domain_realm]
-  .example.com = EXAMPLE.COM
